refactor(Profile): move components related to profile to a `components/profile` folder
