+++ lib/components/item/CommitItem.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 
 import { defaultPalette } from '../../propTypes/palette'
-import ProfilePhoto from '../ProfilePhoto'
+import ProfilePhoto from '../profile/ProfilePhoto'
 import Hash from '../Hash'
 import Datestamp from '../Datestamp'
 import Base from '../Base'
