+++ lib/components/item/ProfileItem.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 
-import ProfilePhoto from '../ProfilePhoto'
+import ProfilePhoto from '../profile/ProfilePhoto'
 import Online from '../Online'
 import { defaultPalette } from '../../propTypes/palette'
 import Base from '../Base'
