+++ lib/components/profile/NoProfile.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import PageHeader from './PageHeader'
-import ReadOnlyHeader from './ReadOnlyHeader'
+import PageHeader from '../PageHeader'
+import ReadOnlyHeader from '../ReadOnlyHeader'
 
-import { defaultPalette } from '../propTypes/palette'
-import Base from './Base'
+import { defaultPalette } from '../../propTypes/palette'
+import Base from '../Base'
 
 export default class NoProfile extends Base {
   constructor (props) {
