+++ lib/components/profile/Profile.js
@@ -1,21 +1,21 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import DatasetRefProps from '../propTypes/datasetRefProps'
-import ProfileProps from '../propTypes/profile'
+import DatasetRefProps from '../../propTypes/datasetRefProps'
+import ProfileProps from '../../propTypes/profile'
 
 import ProfileHeader from './ProfileHeader'
-import TabPanel from './TabPanel'
-import List from './List'
-import DatasetItem from './item/DatasetItem'
-import Spinner from './Spinner'
-import NameDataset from './NameDataset'
+import TabPanel from '../TabPanel'
+import List from '../List'
+import DatasetItem from '../item/DatasetItem'
+import Spinner from '../Spinner'
+import NameDataset from '../NameDataset'
 import NoProfile from './NoProfile'
 import ProfileSidebar from './ProfileSidebar'
-import ProfileEditorContainer from '../containers/ProfileEditor'
-import ReadOnly from './ReadOnly'
+import ProfileEditorContainer from '../../containers/ProfileEditor'
+import ReadOnly from '../ReadOnly'
 
-import Base from './Base'
+import Base from '../Base'
 
 const ADD_PROFILE_DATASET_MODAL = 'ADD_PROFILE_DATASET_MODAL'
 const EDIT_PROFILE_MODAL = 'EDIT_PROFILE_MODAL'
