+++ lib/components/profile/ProfileEditor.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Base from './Base'
-import Spinner from './Spinner'
-import ProfileForm from './form/ProfileForm'
-import Button from './Button'
-import { Palette, defaultPalette } from '../propTypes/palette'
-import ReadOnly from './ReadOnly'
+import Base from '../Base'
+import Spinner from '../Spinner'
+import ProfileForm from '../form/ProfileForm'
+import Button from '../Button'
+import { Palette, defaultPalette } from '../../propTypes/palette'
+import ReadOnly from '../ReadOnly'
 
-import { SET_SESSION_PROFILE_SUCCESS } from '../constants/session'
+import { SET_SESSION_PROFILE_SUCCESS } from '../../constants/session'
 
 export default class ProfileEditor extends Base {
   constructor (props) {
