+++ lib/components/profile/ProfileHeader.js
@@ -3,11 +3,11 @@ import React from 'react'
 import PropTypes from 'prop-types'
 
 import ProfilePhoto from './ProfilePhoto'
-import Online from './Online'
-import Hash from './Hash'
+import Online from '../Online'
+import Hash from '../Hash'
 
-import { defaultPalette } from '../propTypes/palette'
-import Base from './Base'
+import { defaultPalette } from '../../propTypes/palette'
+import Base from '../Base'
 
 export default class ProfileHeader extends Base {
   constructor (props) {
