+++ lib/components/profile/ProfilePhoto.js
@@ -2,8 +2,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import { defaultPalette } from '../propTypes/palette'
-import Base from './Base'
+import { defaultPalette } from '../../propTypes/palette'
+import Base from '../Base'
 
 export default class ProfilePhoto extends Base {
   template (css) {
