+++ lib/components/profile/ProfileSidebar.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Base from './Base'
-import Button from './Button'
+import Base from '../Base'
+import Button from '../Button'
 
 export default class ProfileSidebar extends Base {
   template (css) {
