+++ lib/containers/Profile.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 
-import Profile from '../components/Profile'
+import Profile from '../components/profile/Profile'
 import { showModal, hideModal } from '../actions/app'
 import { loadProfileById, loadProfileByName } from '../actions/profiles'
 import { setProfilePoster, setProfilePhoto, loadSessionProfile } from '../actions/session'
