+++ lib/containers/ProfileEditor.js
@@ -4,7 +4,7 @@ import { loadSessionProfile, saveSessionProfile, updateSession, createLocalSessi
 import { selectSessionProfile, selectLocalProfile } from '../selectors/session'
 import { hideModal } from '../actions/app'
 
-import ProfileEditor from '../components/ProfileEditor'
+import ProfileEditor from '../components/profile/ProfileEditor'
 
 const ProfileEditorContainer = connect(
   (state, ownProps) => {
